Variable S/N ratio used
I have implimented a method to use a non-uniform S/N. The currectly used value for each galaxy is stored in the galaxies.txt file.
NB: python routines should not need updating as they have specified the columns to use (and they don't make use of knowing  the S/N).

 On branch master
 Your branch is up-to-date with 'origin/master'.

 Changes to be committed:
	modified:   binning_spaxels.pro
	modified:   errors.pro
	modified:   full_analysis.pro
	modified:   gandalf_VIMOS.pro
	modified:   mcmc.pro