EoDC: Added new vel normalisation method and isophotos plotting options.
Added ability to plot linear isophotos instead of magnitude with keyword 'flux_type' in plot_velfield_nointerp.py

Update plot_results.py to reflect the above change.
Also added the option normalise the vel field to the mean of the 5 bins with the highest stellar LOSVD, with the 'norm' keyword.

Updated full_analysis.py to reflect the above.

General use of other routines.

 On branch master
 Your branch is up-to-date with 'origin/master'.

 Changes to be committed:
	modified:   collectResults.sh
	modified:   create_params_file.pro
	modified:   full_analysis.pro
	modified:   full_analysis2.py
	modified:   plot_results.py
	modified:   plot_velfield_nointerp.py