EoDC: New routines
plot_results2.py is for plotting the normal outputs of plot_results.py on a gridof subplots.

plot_velfield_nointerp2.py is to handle the plots using on the ax objects so that it can be used by plot_results2.py.

plot_velfield_nointerp.py: sorted orinetation issues. Now have E as left (and increasing RA).

 On branch master
 Changes to be committed:
	modified:   full_analysis2.py
	modified:   plot_results.py
	new file:   plot_results2.py
	modified:   plot_velfield_nointerp.py
	new file:   plot_velfield_nointerp2.py