EoDC: errors3.py: multiplicative not addative continuum
errors3.py: counterpart to errors2.py, but is for stellar population
work, which apparently needs multiplicative polynomials rather than
addative.

checkcomp.py: added the ability to override which platform it is on:
this is due to glamdring nodes not knowing they are part of glamdring.

pickler.py: added 'opt' keyword. This can be 'kin' (default) or 'pop'.
It controls whether it uses the errors2 or errors3 outputs. 'pop' will
save to a different file.

plot_absorption.py and stellar_pop.py: altered to reflect new pickled
file.

Bin.py: add mpweight array and edited calculation of uncovolved_spectrum
to use multiplicative polynomials if present.

full_analysis2.py: added try statements to allow galaxies to fail and
not stop the whole thing running.

 On branch master
 Changes to be committed:
	modified:   Bin.py
	modified:   absorption.py
	modified:   checkcomp.py
	modified:   create_params_file.py
	new file:   errors3.py
	modified:   full_analysis2.py
	modified:   pickler.py
	modified:   plot_absorption.py
	modified:   stellar_pop.py