EoDC: MAJOR fix in stellar_pop.py
stellar_pop.py: found mistake in chi2 calc. Also have normalised chi2
(for each bin) to the min value for that bin in order to have workable
exponetials. NB: this may alter the weighting of the fits...

binning_spaxels.py: changed noise measurement from rms of varience to
median of varience. If rms is used it should be be
rms/sqrt(len(spectrum)), however this would mean that the noise would
reduce for a longer spectrum - which I don't think is right.

General use of full_analysis[2].py

 On branch master
 Changes to be committed:
	modified:   binning_spaxels.py
	modified:   full_analysis.py
	modified:   full_analysis2.py
	modified:   stellar_pop.py