Minor improvements to binning_spaxels.py (and others)
all: increased the space for vel in galaxies.txt by one space.

binning_spaxels: Improved such that it copes with galaxy not being in
the galaxies.txt file better - it will now add an entry if it is
missing.

ssh2glamdring: added opt keyword to allow for pop and kin voronoi files.

 On branch master
 Changes to be committed:
	modified:   binning_spaxels.py
	modified:   kinematics.py
	modified:   mcmc.py
	modified:   ssh2glamdring.sh