Load template files from Data partition instead of home directory
There seems to be some sort of memory issue with loading large numbers of files,
using any method, in python from the home directory. I have solved this by
importing from a location where the files are duplicated in the Data partition.
An if statement means that this only effects my uni computer: the files are still
loaded from ~/models/miles_library/ on other devices.

 On branch master
 Changes to be committed:
	modified:   Bin.py