EoDC: Futher work to Rampazzo apature methods
Rampazzo.py: futher work to the aperture method as well as adding the
in_ellipse method (this is a faster method than funccontains as it does
not need to evaulate func, only is it with the limits.
I am not convinced by the self.r line, however the paper does not
ellaborate clearly enough about this: I have simply taken the mean in
the spectral direction.

compare_absorption.py: Futher work to the Rampazzo section of 'run'
method.

tools.py: moving into this repo, such that it has a github backup.
tools.py is a collection of useful short routines and untilities.

 On branch master
 Changes to be committed:
	modified:   Rampazzo.py
	modified:   compare_absorption.py
	new file:   tools.py