EoDC: Improving calibrations to LICK/IDS system
absorption.py: reduce resolution of unc_spec or conv_spec such that they
match. Also reduce resoltion of spec to LICK resolution. Not sure what
the dispersion is that I should be passing to spectools.py - Should it
be the spectral sampling or the new reduced resolution?
Also need to impliment a wavelength dependant resolution as it is for
the Lick spectra.

Rampazzo.py: updated the definitions for self.r in gradient2 and
apperture methods. It is definately not correct at the moment...

compare_absorption.py: changed to use my measured values for <r_l> from
Rampazzo.py for plots.

 On branch master
 Changes to be committed:
	modified:   Rampazzo.py
	modified:   absorption.py
	modified:   compare_absorption.py