EoDC: Minor alterations
binning_spaxels.py: For some reason the section that saves to file was
not working correctly (not sure how long this has been going on for...)

check.sh: routine was not checking for the final bin - it is now.

errors2.py: improved layout slightly and improved responce from
remove_anomolies routine.

errors3.py: improved layout slightly.

 On branch master
 Changes to be committed:
	modified:   binning_spaxels.py
	modified:   check.sh
	modified:   errors2.py
	modified:   errors3.py