EoDC: Improvements for working on Glamdring
checkcomp.py: method to detect if on a glamdring node using the contents
of the file /etc/clustername. Should not need to use the override method
any longer.

pop.py: Corrections for running on glamdring - still not working though.
Seems to be some issue with importing emcee (Jonathan thinks that is it
detecting a MPI enviorment because of multirun.sh)

 On branch master
 Changes to be committed:
	modified:   checkcomp.py
	modified:   pop.py