EoDC3: Added offset to radio and CO overlays
errors2.py: used get_dataCubeDirctory routine to set an offset to align
the contour plots and VIMOS maps. VLT will not have very accurate
pointing measurements, where as VLA and ALMA get very accurate
pointings. Currently the given offset is for VLA/ALMA -> VLT, which
means that the axis labels may not be as accurate as possible. This
could be switched for the papers/thesis...?

plot_results.py: using and applying the offsets given in errors2.py.

 On branch master
 Changes to be committed:
	modified:   errors2.py
	modified:   plot_results.py