Added ability to place the origin on center of galaxy
plot_velfield_nointerp.py: add center keyword to allow for axis to have
origin on center of galaxy.
This is a commit prior to major overhall of plotting routine. I intend
to plot the image using the coordingates and use twin axis to should
tickmarkers as disired. This should make over plotting other images much
easier.

 On branch master
 Changes to be committed:
	modified:   plot_velfield_nointerp.py